extract usage breakdown into its own component
